Python/Projects/Testing/requirements.txt
# -*- coding: utf-8 -*-
--index-url https://pypi.org/simple
boto3~=1.15
moto~=1.3.16
pandas>=1.1
pytest~=6.2
pytest-mock~=3.6
pyyaml~=5.4
requests~=2.25
